Use our own LCG.
The rand function is thread safe on Linux. But this goes at the
cost of acquiring a lock each time the function is called,
which is not a good idea if it called millions of times
per second.

We now use our own LCG. We avoid contention by giving every
thread its own seed. The seeds are 2^48 steps apart.